Solved #13: Exception when saving a Volunteering Event in the persistence layer

+++ fixmycity/src/main/java/com/ispw/fixmycity/logic/dao/CommunityReportDAO.java
@@ -10,7 +10,7 @@ public interface CommunityReportDAO {
 
 	public CommunityReport save(CommunityReport entity);
 
-	public CommunityReport findByPrimaryKey(Long id);
+	public CommunityReport findByPrimaryKey(Integer id);
 
 	public List<CommunityReportBean> findAll();
